Reasy as content script
This could be faster (although in Chrome there's no way for bare eyes
to notice) as the content script is executed only once at page load;
context menu clicks only invokes ready-to-execute function. However, the
drawback is that content scripts only execute after page is ready; that
means user has to wait until the page is fully loaded to read something!

That's stands so until I find a workaround